Improvements to build script

+++ scripts/buildandcopy.sh
@@ -1,5 +1,27 @@
 #!/bin/bash
 set -e
+PYTHON_CMD="python3"
+
+# Basic flag arg parsing
+MANUAL=0
+DRY=0
+SKIP_WEBPACK_BUILD=0
+for i in $@; do
+    case $i in
+        "--manual")
+        MANUAL=1
+        echo "Doing manual deploy"
+        ;;
+        "--dry")
+        DRY=1
+        echo "Doing dry run"
+        ;;
+        "--skip-wp")
+        SKIP_WEBPACK_BUILD=1
+        ;;
+    esac
+done
+
 
 if ! [[ -z "$DISABLE_COLLECTSTATIC" ]]; then
     echo "Collectstatic disabled"
@@ -29,12 +51,14 @@ fi
 mkdir -p ./python/aristotle-metadata-registry/aristotle_mdr/manifests
 
 cd assets
-echo "Running webpack build..."
-# Remove stats if it exists
-rm -f ./dist/webpack-stats.json
-npm install
-npm run build
-echo "Webpack build complete!"
+if [[ $SKIP_WEBPACK_BUILD -ne 1 ]]; then
+    echo "Running webpack build..."
+    # Remove stats if it exists
+    rm -f ./dist/webpack-stats.json
+    npm install
+    npm run build
+    echo "Webpack build complete!"
+fi
 cd ..
 
 # If aws command not avaliable
@@ -45,7 +69,7 @@ fi
 
 echo "Collecting bundle static..."
 
-if [[ "$1" == "--dry" ]]; then
+if [[ $DRY -eq 1 ]]; then
     aws s3 cp ./assets/dist/bundles s3://$STORAGE_BUCKET_NAME/bundles --recursive --dryrun
 else
     aws s3 cp ./assets/dist/bundles s3://$STORAGE_BUCKET_NAME/bundles --recursive
@@ -53,11 +77,15 @@ fi
 
 cp ./assets/dist/webpack-stats.json ./python/aristotle-metadata-registry/aristotle_mdr/manifests
 
-if [[ "$1" == "--manual" ]]; then
+if [[ $MANUAL -eq 1 ]]; then
     echo "Doing a manual deploy to s3..."
     rm -r ./dist
-    python setup.py bdist_wheel
-    aws s3 cp ./dist s3://aristotle-pypi-bucket-1kyswb3cn1pa1 --recursive 
+    $PYTHON_CMD setup.py bdist_wheel
+    if [[ $DRY -eq 1 ]]; then
+        aws s3 cp ./dist s3://aristotle-pypi-bucket-1kyswb3cn1pa1 --recursive --dry
+    else
+        aws s3 cp ./dist s3://aristotle-pypi-bucket-1kyswb3cn1pa1 --recursive 
+    fi
 fi
 
 echo "Done!"
